code cleanup - resolve crash for settings view
